dodicesima commit

+++ db/migrate/20170914090844_add_price_to_owners.rb
@@ -1,5 +1,5 @@
 class AddPriceToOwners < ActiveRecord::Migration
   def change
-    add_column :owners, :price, :integer
+    add_column :owners, :price, :float
   end
 end
